Add Bullet and Awan

+++ NewBuffer/engine.c
@@ -22,7 +22,7 @@ int fbfd = 0;
 
 void initFbp(){
     fbp = (char *)mmap(0, screensize, PROT_READ | PROT_WRITE, MAP_SHARED, fbfd, 0);
-    if ((int)fbp == -1) {
+    if ((long)fbp == -1) {
         perror("Error: failed to map framebuffer device to memory");
         exit(4);
     }
